Add files via upload
An attempt to fix promotion logic, this version does have some other improvements, but overall, it still doesn't work 100% the way it should. When a server is promoted the client can only communicate with the server that is assigned as primary. 

I think I might have to make it so client can only communicate with port 8090, but IDK I will see tomorrow when I have some time.